Moved draw functions to render.py + Test cases for movement and projectile.
