+++ src/main.py
@@ -1,12 +1,11 @@
 # main menu
 import pygame
 import asyncio
-import math
 import sys
 
 from .player import Player
 from .enemy import Enemy
-from .render import draw_objects, apply_brightness
+from .render import draw_objects, apply_brightness, draw_credits, render_menu_screen, render_inventory_screen, render_game_over_screen
 from .camera import Camera          # Added for camera - Meheraj
 from .background import SpaceBackground # Added for parallax background - Meheraj
 from .world import World
@@ -23,114 +22,6 @@ GAME = "game"
 GAME_OVER = "game_over"
 CREDITS = "Credits"
 INVENTORY = "inventory"
-
-def draw_menu(win, screen_width, screen_height, truck_img, float_time):
-
-    title_font = pygame.font.SysFont(None, 90)
-    small_font = pygame.font.SysFont(None, 32)
-    btn_font = pygame.font.SysFont(None, 48)
-
-    # Title
-    title = title_font.render("STATION FALL", True, (255, 255, 255))
-    win.blit(title, title.get_rect(center=(screen_width // 2, screen_height // 2 - 180)))
-
-    hint = small_font.render("Press ENTER to start", True, (200, 200, 200))
-    win.blit(hint, hint.get_rect(center=(screen_width // 2, screen_height // 2 - 120)))
-
-    #ship
-    float_offset = math.sin(float_time) * 15
-    truck_rect = truck_img.get_rect()
-    truck_rect.center = (screen_width // 2, screen_height // 2 + float_offset)
-    win.blit(truck_img, truck_rect)
-
-    # Buttons
-    btn_w, btn_h = 260, 60
-    start_rect = pygame.Rect(0, 0, btn_w, btn_h)
-    start_rect.center = (screen_width // 2, screen_height // 2 +120)
-
-    credits_rect = pygame.Rect(0, 0, btn_w, btn_h)
-    credits_rect.center = (screen_width // 2, screen_height // 2 + 210)
-
-    quit_rect = pygame.Rect(0, 0, btn_w, btn_h)
-    quit_rect.center = (screen_width // 2, screen_height // 2 + 300)
-
-    mx, my = pygame.mouse.get_pos()
-
-    def draw_button(rect, text):
-        hover = rect.collidepoint(mx, my)
-        color = (220, 220, 220) if hover else (170, 170, 170)
-        pygame.draw.rect(win, color, rect, border_radius=12)
-        pygame.draw.rect(win, (40, 40, 40), rect, 3, border_radius=12)
-
-        label = btn_font.render(text, True, (0, 0, 0))
-        win.blit(label, label.get_rect(center=rect.center))
-
-    draw_button(start_rect, "Start")
-    draw_button(credits_rect, "Credits")
-    draw_button(quit_rect, "Quit")
-
-    return start_rect, quit_rect, credits_rect
-
-# Game over screen
-def draw_game_over(win, screen_width, screen_height, truck_img, float_time):
-    font = pygame.font.SysFont(None, 90)
-    small_font = pygame.font.SysFont(None, 32)
-
-    text = font.render("GAME OVER", True, (255, 0, 0))
-    hint = small_font.render("Press ENTER to restart or ESC to quit", True, (255, 255, 255))
-
-    win.blit(text, text.get_rect(center=(screen_width//2, screen_height//2 - 50)))
-    win.blit(hint, hint.get_rect(center=(screen_width//2, screen_height//2 + 50)))
-
-    # Floating truck (top of screen)
-    float_offset = math.sin(float_time) * 15
-    truck_rect = truck_img.get_rect()
-    truck_rect.midtop = (screen_width // 2, 200 + float_offset)
-    win.blit(truck_img, truck_rect)
-
-# Credits screen - Wil
-def draw_credits(win, screen_width, screen_height, background, float_time, menu_camera_x, menu_camera_y):
-    background.update_and_draw(win, (menu_camera_x, menu_camera_y))
-
-    title_font = pygame.font.SysFont(None, 72)
-    name_font = pygame.font.SysFont(None, 42)
-    role_font = pygame.font.SysFont(None, 30)
-    hint_font = pygame.font.SysFont(None, 28)
-
-    # Title
-    title = title_font.render("CREDITS", True, (255, 255, 255))
-    win.blit(title, title.get_rect(center=(screen_width // 2, 80)))
-
-    # Divider line
-    pygame.draw.line(win, (100, 100, 255), (screen_width // 2 - 200, 120), (screen_width // 2 + 200, 120), 2)
-
-    team = [
-        ("Wil Nahra",            "| Developer | Sprite Creation |"),
-        ("Simon Halaszi",        "| Developer |"),
-        ("Loy Ngo",              "| Developer |"),
-        ("Mark",                 "| Developer |"),
-        ("Rowan",                "| Developer |"),
-        ("Sebastian Bentancourt","| Developer |"),
-        ("Yusairah Haque",       "| Developer |"),
-        ("Zachary Evans",        "| Developer |"),
-    ]
-
-    start_y = 170
-    spacing = 90
-
-    for i, (name, role) in enumerate(team):
-        y = start_y + i * spacing
-        float_offset = math.sin(float_time + i * 0.4) * 4
-
-        name_surf = name_font.render(name, True, (220, 220, 255))
-        role_surf = role_font.render(role, True, (140, 140, 200))
-
-        win.blit(name_surf, name_surf.get_rect(center=(screen_width // 2, y + float_offset)))
-        win.blit(role_surf, role_surf.get_rect(center=(screen_width // 2, y + 32 + float_offset)))
-
-    # ESC hint
-    hint = hint_font.render("Press ESC to return", True, (160, 160, 160))
-    win.blit(hint, hint.get_rect(center=(screen_width // 2, screen_height - 40)))
 
 async def main(): 
     pygame.init()
@@ -275,10 +166,9 @@ async def main():
         win.fill((0, 0, 0))
 
         if state == MENU:
-            background.update_and_draw(win, (menu_camera_x, menu_camera_y))
-            start_rect, quit_rect, credits_rect = draw_menu(
-                win, screen_width, screen_height, truck_img, float_time
-            )
+            start_rect, quit_rect, credits_rect = render_menu_screen(
+            win, screen_width, screen_height, background, menu_camera_x, 
+            menu_camera_y, truck_img, float_time)
 
         elif state == GAME:
             keys = pygame.key.get_pressed()
@@ -327,13 +217,11 @@ async def main():
                 draw_objects(win, player, enemies, bullets, world.walls, camera, background, coins)
 
         elif state == INVENTORY:
-            background.update_and_draw(win, (camera.camera.x, camera.camera.y))
-            draw_objects(win, player, enemies, bullets, world.walls, camera, background, coins)
-            inventory_ui.draw(win, player.money)
-
+            render_inventory_screen(win, player, enemies, bullets, world.walls,
+                                    camera, background, coins, inventory_ui)
         elif state == GAME_OVER:
-            background.update_and_draw(win, (menu_camera_x, menu_camera_y))
-            draw_game_over(win, screen_width, screen_height, gameOver_img, float_time)
+            render_game_over_screen(win, screen_width, screen_height, background, 
+                                    menu_camera_x, menu_camera_y, gameOver_img, float_time)
 
         elif state == CREDITS:
             draw_credits(win, screen_width, screen_height, background, float_time, menu_camera_x, menu_camera_y)
